+++ BankingSystem/improvement.txt
@@ -1,0 +1,31 @@
+1. **Code Refactoring & Cleanup**  
+   • Rename methods and variables to follow Java naming conventions (e.g., change `displayaccount()` to `displayAccount()`).  
+   • Separate the user interface (UI) logic (input/output) from the core business logic. This will make future changes easier to manage.
+
+2. **Enhance Data Structures & Input Validation**  
+   • Replace the fixed-size array with a dynamic structure like an `ArrayList` for storing accounts.  
+   • Add thorough input validation (e.g., checking for non-numeric input, ensuring the initial balance meets the minimum requirement, verifying positive deposit/withdrawal amounts).
+
+3. **Implement Robust Error Handling**  
+   • Integrate exception handling to manage invalid user input and runtime errors gracefully.  
+   • Consider logging error messages for easier debugging and maintenance.
+
+4. **Introduce Logging**  
+   • Add a logging mechanism (using Java’s built-in logging utilities or a third-party library) to track system events and transactions. This will be useful for both debugging and auditing.
+
+5. **Persistent Data Storage**  
+   • Shift from in-memory storage to file-based or database storage so that account data and transaction histories persist between sessions.
+
+6. **Security Enhancements**  
+   • Implement basic authentication (e.g., a login system with a PIN or password) to restrict access to account operations.  
+   • Encrypt sensitive data where necessary to safeguard user information.
+
+7. **Additional Feature: Fund Transfer**  
+   • Add functionality for transferring funds between accounts, including checks for account validity and sufficient funds.
+
+8. **Additional Feature: Transaction History**  
+   • Implement a transaction history log that records deposits, withdrawals, and transfers along with timestamps. This can help with auditing and give users insight into their account activity.
+
+9. **Further Enhancements (Interest, Fees, and UI Improvements)**  
+   • For savings accounts, add interest calculations, and for current accounts, consider fee deductions.  
+   • Revamp the UI with clearer instructions and possibly explore a GUI for a more interactive experience.
